refactor: move getNeighbors function to own file
